feat(drag-drop): move into cdk
Moves the `drag-drop` module from `cdk-experimental` to `cdk`.
