+++ src/cdk/drag-drop/drag-drop-registry.ts
