+++ src/cdk/drag-drop/drag-placeholder.ts
