+++ src/e2e-app/e2e-app-module.ts
@@ -1,5 +1,5 @@
 import {DialogModule} from '@angular/cdk-experimental/dialog';
-import {DragDropModule} from '@angular/cdk-experimental/drag-drop';
+import {DragDropModule} from '@angular/cdk/drag-drop';
 import {ScrollingModule as ExperimentalScrollingModule} from '@angular/cdk-experimental/scrolling';
 import {FullscreenOverlayContainer, OverlayContainer} from '@angular/cdk/overlay';
 import {ScrollingModule} from '@angular/cdk/scrolling';
